Add insurance details routing
