+++ apps/store/src/features/memberArea/pages/InsuranceDetailsPage.tsx
@@ -1,0 +1,11 @@
+import { NextPageWithLayout } from 'next'
+import { useMemberAreaMemberInfoQuery } from '@/services/apollo/generated'
+import { LayoutWithMenu } from '../components/LayoutWithMenu'
+import { InsuranceDetailsSection } from '../InsuranceSection/InsuranceDetailsSection'
+
+export const InsuranceDetailsPage: NextPageWithLayout = () => {
+  const { loading } = useMemberAreaMemberInfoQuery()
+  return <>{loading ? 'Loading...' : <InsuranceDetailsSection />}</>
+}
+
+InsuranceDetailsPage.getLayout = (children) => <LayoutWithMenu>{children}</LayoutWithMenu>
