+++ apps/store/src/pages/member/insurances/[id].tsx
@@ -1,0 +1,5 @@
+import { InsuranceDetailsPage } from '@/features/memberArea/pages/InsuranceDetailsPage'
+import { protectedPageServerSideProps } from '@/features/memberArea/pages/protectedPageServerSideProps'
+
+export const getServerSideProps = protectedPageServerSideProps
+export default InsuranceDetailsPage
